fix(kotlin): AtomicLong.addAndGet takes Long — chunk length is Int

+++ app/src/main/java/io/github/xixka/qbittorrent/util/ApkDownloader.kt
@@ -127,7 +127,7 @@ object ApkDownloader {
                 while (true) {
                     try {
                         downloadPart(url, dest, start + partDone.get(), end, partDone) { n ->
-                            done.addAndGet(n)
+                            done.addAndGet(n.toLong())
                             onProgress(Progress(done.get(), total))
                         }
                         return@async
